Refactor: Unfinished make authorize

+++ app/Http/Controllers/PosisiJabatanController.php
@@ -65,6 +65,8 @@ class PosisiJabatanController extends Controller
      */
     public function store(PosisiJabatanRequest $request)
     {
+        $this->authorize('admin-officer');
+
         $this->posisijabatanRepository->store($request);
         return back()->with('success', 'Berhasil menambah Posisi Jabatan');
     }
@@ -82,6 +84,8 @@ class PosisiJabatanController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('admin-officer');
+
         $posisiJabatan = $this->posisijabatanRepository->edit($id);
     }
 
@@ -90,6 +94,8 @@ class PosisiJabatanController extends Controller
      */
     public function update(PosisiJabatanRequest $request, string $id)
     {
+        $this->authorize('admin-officer');
+
         $this->posisijabatanRepository->update($id, $request);
         return back()->with('success', 'Berhasil meng-edit Posisi Jabatan');
     }
@@ -99,6 +105,8 @@ class PosisiJabatanController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('admin-officer');
+
         $encryptId = Crypt::decryptString($id);
         $this->posisijabatanRepository->destroy($encryptId);
         return back()->with('success', 'Berhasil menghapus Posisi Jabatan');
